# Micro-optimization: Changed switch to ifs, added benchmarks